Вывод постов(первая стадия)
- Модифицирована модель Post, добавлены новые методы которые позволяют
с помощью Carbon модифицировать немного created_at, для его красивого вывода
Также настроен корректный вывод картинок (ещё один метод)

- Работа с blade и вывод циклом постов (пока без логики перехода на single страничку и без логики просмотров)

+++ app/Http/Controllers/AdminPanel/CategoryController.php
@@ -62,7 +62,10 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        if($category->posts()->count()){
+            return redirect()->route('category.index')->with('error','Ошибка! У категории есть записи!');
+        }
         $category->delete($category);
-        return redirect()->route('category.index')->with('success','Категория успешно редактирована!');
+        return redirect()->route('category.index')->with('success','Категория успешно удалена!');
     }
 }
